requirements.txt
aioredis==1.2.0
appdirs==1.4.3
apturl==0.5.2
asgiref==2.3.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
command-not-found==0.3
configparser==3.7.3
constantly==15.1.0
cryptography==2.8
cssselect==1.0.3
cupshelpers==1.0
daphne==2.2.5
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
djangorestframework-bulk==0.2.1
duplicity==0.8.12.0
entrypoints==0.3
fake-useragent==0.1.11
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
google==2.0.2
gunicorn==20.0.4
hiredis==1.0.0
httplib2==0.14.0
hyperlink==18.0.0
idna==2.8
importlib-metadata==1.6.1
incremental==17.5.0
isort==4.3.21
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.3.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
msgpack==0.6.1
netifaces==0.10.4
oauthlib==3.0.1
olefile==0.46
paramiko==2.6.0
parse==1.12.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pyee==6.0.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyppeteer==0.0.25
pyquery==1.4.0
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2018.9
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.21.0
requests-html==0.10.0
requests-oauthlib==1.2.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.1
stevedore==2.0.0
systemd-python==234
toml==0.10.1
tqdm==4.31.1
tweepy==3.7.0
Twisted==18.9.0
txaio==18.8.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.7
virtualenv==20.0.23
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.20.0
wadllib==1.3.3
websockets==7.0
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.6.0
